Extract arrow path logic into Arrow module
This keeps the view thin, all my markup in the template, and all my
elixir heavy lifting in the Chukinas.Dreadnought dir/namespace. I'm not
sure yet whether I want view helpers like Arrow included amongst the
core game logic e.g. Dreadnought.Deck, but I can always change that
later.
